move add article functionality to a separate file
